stashing my changes to investigate a bug
refers to #7
