+++ VAM Evolutionary Character Creation.py
@@ -11,48 +11,20 @@ import sys
 import json
 import tkinter as tk
 
+import ecc_utility
 import ecc_gui
 import ecc_logic
 
 
 BG_COLOR = "#F9F9F9"
-DATA_PATH = "data"
-SETTINGS_FILENAME = "settings.json"
 ICON_FILENAME = "VAM Evolutionary Character Creation.ico"
-
-
-def save_settings(settings):
-    """ Saves the settings as a json file to DATA_PATH/SETTINGS_FILENAME """
-    if getattr(sys, 'frozen', False):
-        dir_path = os.path.dirname(sys.executable)
-    elif __file__:
-        dir_path = os.path.dirname(os.path.realpath(__file__))
-    filename = os.path.join(dir_path, DATA_PATH, SETTINGS_FILENAME)
-    with open(filename, 'w') as json_file:
-        print("Writing settings to:", filename)
-        json.dump(settings, json_file, indent=3)
-
-
-def load_settings():
-    """ Fills settings with the settings in the DATA_PATH/SETTINGS_FILENAME json_file """
-    settings = {}
-    if getattr(sys, 'frozen', False):
-        dir_path = os.path.dirname(sys.executable)
-    elif __file__:
-        dir_path = os.path.dirname(os.path.realpath(__file__))
-    filename = os.path.join(dir_path, DATA_PATH, SETTINGS_FILENAME)
-    if os.path.isfile(filename):
-        with open(filename) as json_file:
-            print("Reading settings from:", filename)
-            settings = json.load(json_file)
-    return settings
 
 
 def main():
     '''
     create the business logic and the main window and launch them
     '''
-    settings = load_settings()
+    settings = ecc_utility.load_settings()
 
     # todo: convert Logic into class later
 
@@ -61,14 +33,14 @@ def main():
     main_window = tk.Tk()
     main_window.configure(bg=BG_COLOR)
     main_window.option_add("*font", "Calibri 9")
-    main_window.iconbitmap(os.path.join(DATA_PATH, ICON_FILENAME))
+    main_window.iconbitmap(os.path.join(ecc_utility.DATA_PATH, ICON_FILENAME))
 
     app = ecc_gui.AppWindow(settings, generator)
 
     app.initialize()
     main_window.mainloop()
 
-    save_settings(settings)
+    ecc_utility.save_settings(settings)
 
 
 if __name__ == '__main__':
